Refactors explorers into multiple files, comments and cleanup.
