+++ mathjax3-ts/a11y/explorer/Region.ts
@@ -1,6 +1,6 @@
 /*************************************************************
  *
- *  Copyright (c) 2009-2018 The MathJax Consortium
+ *  Copyright (c) 2009-2019 The MathJax Consortium
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
